requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.2
django-widget-tweaks==1.5.0
mysqlclient==2.2.7
pillow==11.2.1
python-decouple==3.8
qrcode==8.1
reportlab==4.4.0
sqlparse==0.5.3
tzdata==2025.2
